+++ src/dq_suite/__init__.py
@@ -2,7 +2,7 @@
 
 from .common import ValidationSettings
 from .df_checker import run
-from .input_helpers import schema_to_json_string
+from .validation_input import schema_to_json_string
 
 # Use __all__ to let developers know what is part of the public API.
 __all__ = ["schema_to_json_string", "run", "ValidationSettings"]
